+++ sdk/keyvault/keyvault-keys/samples/javascript/helloWorld.js
@@ -46,9 +46,14 @@ async function main() {
   const updatedKey = await client.updateKey(keyName, result.properties.version, { enabled: false });
   console.log("updated key: ", updatedKey);
 
-  await client.beginDeleteKey(keyName).pollUntilDone();
-  await client.beginDeleteKey(ecKeyName).pollUntilDone();
-  await client.beginDeleteKey(rsaKeyName).pollUntilDone();
+  let poller = await client.beginDeleteKey(keyName);
+  await poller.pollUntilDone();
+
+  poller = await client.beginDeleteKey(ecKeyName);
+  await poller.pollUntilDone();
+
+  poller = await client.beginDeleteKey(rsaKeyName);
+  await poller.pollUntilDone();
 }
 
 main().catch((err) => {