+++ sdk/keyvault/keyvault-keys/src/keysModels.ts
@@ -16,6 +16,8 @@ import { DeletionRecoveryLevel } from "./core/models";
  * An interface representing the key client. For internal use.
  */
 export interface KeyClientInterface {
+  recoverDeletedKey(name: string, options?: RequestOptions): Promise<Key>;
+  getKey(name: string, options?: GetKeyOptions): Promise<Key>;
   deleteKey(name: string, options?: coreHttp.RequestOptionsBase): Promise<DeletedKey>;
   getDeletedKey(name: string, options?: RequestOptions): Promise<DeletedKey>;
 }
@@ -101,25 +103,34 @@ export interface KeyProperties extends ParsedKeyVaultEntityIdentifier {
  * @interface
  * An interface representing a deleted key
  */
-export interface DeletedKey extends Key {
+export interface DeletedKey {
   /**
-   * @member {string} [recoveryId] The url of the recovery object, used to
-   * identify and recover the deleted key.
+   * @member {string} [value] The key value.
    */
-  readonly recoveryId?: string;
+  keyMaterial?: JsonWebKey;
   /**
-   * @member {Date} [scheduledPurgeDate] The time when the key is scheduled
-   * to be purged, in UTC
-   * **NOTE: This property will not be serialized. It can only be populated by
-   * the server.**
+   * @member {KeyProperties} [properties] The properties of the key.
    */
-  readonly scheduledPurgeDate?: Date;
-  /**
-   * @member {Date} [deletedDate] The time when the key was deleted, in UTC
-   * **NOTE: This property will not be serialized. It can only be populated by
-   * the server.**
-   */
-  readonly deletedDate?: Date;
+  properties: KeyProperties & {
+    /**
+     * @member {string} [recoveryId] The url of the recovery object, used to
+     * identify and recover the deleted key.
+     */
+    readonly recoveryId?: string;
+    /**
+     * @member {Date} [scheduledPurgeDate] The time when the key is scheduled
+     * to be purged, in UTC
+     * **NOTE: This property will not be serialized. It can only be populated by
+     * the server.**
+     */
+    readonly scheduledPurgeDate?: Date;
+    /**
+     * @member {Date} [deletedDate] The time when the key was deleted, in UTC
+     * **NOTE: This property will not be serialized. It can only be populated by
+     * the server.**
+     */
+    readonly deletedDate?: Date;
+  };
 }
 
 /**
@@ -161,7 +172,7 @@ export interface CreateKeyOptions {
  * An interface representing the optional parameters that can be
  * passed to beginDeleteKey
  */
-export interface BeginDeleteKeyOptions {
+export interface KeyPollerOptions {
   /**
    * @member {coreHttp.RequestOptionsBase} [requestOptions] Options for this request
    */