+++ sdk/keyvault/keyvault-keys/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
 ## 4.0.0-preview.9 (2019-10-31)
-- deleteKey is now out of the public API. Use beginDeleteKey instead. beginDeleteKey returns a Long Running Operation Poller (from our package @azure/core-lro).
+- deleteKey and recoverDeletedKey are now out of the public API.
+  Use beginDeleteKey and beginRecoverDeletedKey instead.
+  They both return a Poller (from our package @azure/core-lro) that manages the long running operation.
 
 ## 4.0.0-preview.8 (2019-10-09)